revert peek, create_server, move test
